+++ RTClib.h
@@ -22,7 +22,7 @@ public:
     // 32-bit times as seconds since 1/1/2000
     long secondstime() const;   
     // 32-bit times as seconds since 1/1/1970
-    uint32_t unixtime(void) const;
+    uint64_t unixtime(void) const;
 
 protected:
     uint8_t yOff, m, d, hh, mm, ss;
